WIP: python wrapper for dbd module + data pickling
A wrapper to make it *far* easier to wrangle things related to DBD,
find columns that exist in a given build, etc, plus matching changes
to the dbd_to_mysql dump script.

The structures returned by the wrapper are also pickleable, unlike the
structures created by the underlying dbd module (courtesy modgrammar),
so this means we can cache what we've parsed for a speedup of ~50x when
the cache exists.